Sorting works. Just run Sorter by itself to see.
I copied Fibonacci and reworked it from the inside out to make Sorter.
It uses the same Producer and Consumer, but has modifications to fit
this project. SpinBoxWindow was also taken out because it's not longer
necessary.

When you run it, TextEdit will appaear as a window and populate because
Producer will randomly generate 50 int numbers from 1 to 1000, and then
Consumer will sort them and output them line by line.